backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
pillow>=10.0.0
pypdf2>=3.0.0
qrcode[pil]>=7.4.0
pyzbar>=0.1.9
httpx>=0.25.0
pymupdf>=1.23.0